Apply suggestion - Allow adding highscores only after winning a game
